+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/AmProps.java
@@ -29,5 +29,6 @@ public class AmProps {
     public static boolean ALERTMONITOR_KAFKA_ENABLED = false;
     public static String ALERTMONITOR_KAFKA_SERVER = "kafkahost:9092";
     public static String ALERTMONITOR_KAFKA_TOPIC = "alertmonitor_alerts";
+    public static String ALERTMONITOR_PROMETHEUS_ID_LABELS = "cluster, region, monitor";
 
 }