+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/LogFactory.java
@@ -5,6 +5,7 @@ import si.matjazcerkvenik.simplelogger.SimpleLogger;
 public class LogFactory {
 
     private static SimpleLogger logger = null;
+    private static SimpleLogger alertLog = null;
 
     public static SimpleLogger getLogger() {
         if (logger == null) {
@@ -15,5 +16,13 @@ public class LogFactory {
         }
         return logger;
     }
+
+    public static SimpleLogger getAlertLog() {
+        if (alertLog == null) {
+            alertLog = new SimpleLogger("./alerts.log");
+        }
+        return alertLog;
+    }
+
 
 }